+++ Moya/Moya.swift
@@ -1,11 +1,11 @@
 import Foundation
 import Alamofire
 
-/// Block to be executed when a request has completed.
-public typealias MoyaCompletion = (data: NSData?, statusCode: Int?, response: NSURLResponse?, error: ErrorType?) -> ()
-
 /// General-purpose class to store some enums and class funcs.
 public class Moya {
+
+    /// Closure to be executed when a request has completed.
+    public typealias Completion = (data: NSData?, statusCode: Int?, response: NSURLResponse?, error: ErrorType?) -> ()
 
     /// Network activity change notification type.
     public enum NetworkActivityChangeType {
@@ -95,38 +95,48 @@ struct CancellableToken: Cancellable {
 }
 
 /// Request provider class. Requests should be made through this class only.
-public class MoyaProvider<T: MoyaTarget> {
-    /// Closure that defines the endpoints for the provider.
-    public typealias MoyaEndpointsClosure = (T) -> (Endpoint<T>)
-    /// Closure that resolves an Endpoint into an NSURLRequest.
-    public typealias MoyaEndpointResolution = (endpoint: Endpoint<T>) -> (NSURLRequest)
-    public typealias StubClosure = (T -> Moya.StubBehavior)
+public class MoyaProvider<Target: MoyaTarget> {
 
-    public let endpointClosure: MoyaEndpointsClosure
-    public let endpointResolver: MoyaEndpointResolution
-    public let stubBehavior: StubClosure
+    /// Closure that defines the endpoints for the provider.
+    public typealias EndpointClosure = Target -> Endpoint<Target>
+
+    /// Closure that resolves an Endpoint into an NSURLRequest.
+    public typealias RequestClosure = Endpoint<Target> -> NSURLRequest
+
+    /// Closure that decides if/how a request should be stubbed.
+    public typealias StubClosure = Target -> Moya.StubBehavior
+
+
+    public let endpointClosure: EndpointClosure
+    public let requestClosure: RequestClosure
+    public let stubClosure: StubClosure
     public let networkActivityClosure: Moya.NetworkActivityClosure?
     public let manager: Manager
 
     /// Initializes a provider.
-    public init(endpointClosure: MoyaEndpointsClosure = MoyaProvider.DefaultEndpointMapping, endpointResolver: MoyaEndpointResolution = MoyaProvider.DefaultEndpointResolution, stubBehavior: StubClosure = MoyaProvider.NeverStub, networkActivityClosure: Moya.NetworkActivityClosure? = nil, manager: Manager = Alamofire.Manager.sharedInstance) {
+    public init(endpointClosure: EndpointClosure = MoyaProvider.DefaultEndpointMapping,
+        requestClosure: RequestClosure = MoyaProvider.DefaultRequestMapping,
+        stubClosure: StubClosure = MoyaProvider.NeverStub,
+        networkActivityClosure: Moya.NetworkActivityClosure? = nil,
+        manager: Manager = Alamofire.Manager.sharedInstance) {
+
         self.endpointClosure = endpointClosure
-        self.endpointResolver = endpointResolver
-        self.stubBehavior = stubBehavior
+        self.requestClosure = requestClosure
+        self.stubClosure = stubClosure
         self.networkActivityClosure = networkActivityClosure
         self.manager = manager
     }
 
     /// Returns an Endpoint based on the token, method, and parameters by invoking the endpointsClosure.
-    public func endpoint(token: T) -> Endpoint<T> {
+    public func endpoint(token: Target) -> Endpoint<Target> {
         return endpointClosure(token)
     }
 
-    /// Designated request-making method.
-    public func request(token: T, completion: MoyaCompletion) -> Cancellable {
+    /// Designated request-making method. Returns a Cancellable token to cancel the request later.
+    public func request(token: Target, completion: Moya.Completion) -> Cancellable {
         let endpoint = self.endpoint(token)
-        let request = endpointResolver(endpoint: endpoint)
-        let stubBehavior = self.stubBehavior(token)
+        let request = requestClosure(endpoint)
+        let stubBehavior = self.stubClosure(token)
 
         switch stubBehavior {
         case .Never:
@@ -135,36 +145,46 @@ public class MoyaProvider<T: MoyaTarget> {
             return stubRequest(request, completion: completion, endpoint: endpoint, stubBehavior: stubBehavior)
         }
     }
+}
 
-    public class func DefaultEndpointMapping(target: T) -> Endpoint<T> {
+/// Mark: Defaults
+
+public extension MoyaProvider {
+
+    // These functions are default mappings to endpoings and requests.
+
+    public class func DefaultEndpointMapping(target: Target) -> Endpoint<Target> {
         let url = target.baseURL.URLByAppendingPathComponent(target.path).absoluteString
         return Endpoint(URL: url, sampleResponse: .Success(200, {target.sampleData}), method: target.method, parameters: target.parameters)
     }
 
-    @available(*, unavailable, renamed="DefaultEndpointResolution", message="Use #DefaultEndpointResolution method instead")
-    public class func DefaultEnpointResolution(endpoint: Endpoint<T>) -> NSURLRequest {
-        return DefaultEndpointResolution(endpoint)
-    }
-
-    public class func DefaultEndpointResolution(endpoint: Endpoint<T>) -> NSURLRequest {
+    public class func DefaultRequestMapping(endpoint: Endpoint<Target>) -> NSURLRequest {
         return endpoint.urlRequest
     }
+}
 
-    public class func NeverStub(_: T) -> Moya.StubBehavior {
+/// Mark: Stubbing
+
+public extension MoyaProvider {
+
+    // Swift won't let us put the StubBehavior enum inside the provider class, so we'll
+    // at least add some class functions to allow easy access to common stubbing closures.
+
+    public class func NeverStub(_: Target) -> Moya.StubBehavior {
         return .Never
     }
 
-    public class func ImmediatelyStub(_: T) -> Moya.StubBehavior {
+    public class func ImmediatelyStub(_: Target) -> Moya.StubBehavior {
         return .Immediate
     }
 
-    public class func DelayedStub(seconds: NSTimeInterval)(_: T) -> Moya.StubBehavior {
+    public class func DelayedStub(seconds: NSTimeInterval)(_: Target) -> Moya.StubBehavior {
         return .Delayed(seconds: seconds)
     }
 }
 
 private extension MoyaProvider {
-    func sendRequest(request: NSURLRequest, completion: MoyaCompletion) -> CancellableToken {
+    func sendRequest(request: NSURLRequest, completion: Moya.Completion) -> CancellableToken {
 
         networkActivityClosure?(change: .Began)
 
@@ -189,7 +209,7 @@ private extension MoyaProvider {
         }
     }
 
-    func stubRequest(request: NSURLRequest, completion: MoyaCompletion, endpoint: Endpoint<T>, stubBehavior: Moya.StubBehavior) -> CancellableToken {
+    func stubRequest(request: NSURLRequest, completion: Moya.Completion, endpoint: Endpoint<Target>, stubBehavior: Moya.StubBehavior) -> CancellableToken {
         var canceled = false
         let cancellableToken = CancellableToken { canceled = true }