+++ Moya/RxSwift/Moya+RxSwift.swift
@@ -3,18 +3,23 @@ import RxSwift
 import Alamofire
 
 /// Subclass of MoyaProvider that returns Observable instances when requests are made. Much better than using completion closures.
-public class RxMoyaProvider<T where T: MoyaTarget>: MoyaProvider<T> {
+public class RxMoyaProvider<Target where Target: MoyaTarget>: MoyaProvider<Target> {
     /// Current requests that have not completed or errored yet.
     /// Note: Do not access this directly. It is public only for unit-testing purposes (sigh).
-    public var inflightRequests = Dictionary<Endpoint<T>, Observable<MoyaResponse>>()
+    public var inflightRequests = Dictionary<Endpoint<Target>, Observable<MoyaResponse>>()
 
     /// Initializes a reactive provider.
-    override public init(endpointClosure: MoyaEndpointsClosure = MoyaProvider.DefaultEndpointMapping, endpointResolver: MoyaEndpointResolution = MoyaProvider.DefaultEndpointResolution, stubBehavior: StubClosure = MoyaProvider.NeverStub, networkActivityClosure: Moya.NetworkActivityClosure? = nil, manager: Manager = Alamofire.Manager.sharedInstance) {
-        super.init(endpointClosure: endpointClosure, endpointResolver: endpointResolver, stubBehavior: stubBehavior, networkActivityClosure: networkActivityClosure, manager: manager)
+    override public init(endpointClosure: EndpointClosure = MoyaProvider.DefaultEndpointMapping,
+        requestClosure: RequestClosure = MoyaProvider.DefaultRequestMapping,
+        stubClosure: StubClosure = MoyaProvider.NeverStub,
+        networkActivityClosure: Moya.NetworkActivityClosure? = nil,
+        manager: Manager = Alamofire.Manager.sharedInstance) {
+        
+        super.init(endpointClosure: endpointClosure, requestClosure: requestClosure, stubClosure: stubClosure, networkActivityClosure: networkActivityClosure, manager: manager)
     }
 
     /// Designated request-making method.
-    public func request(token: T) -> Observable<MoyaResponse> {
+    public func request(token: Target) -> Observable<MoyaResponse> {
         let endpoint = self.endpoint(token)
 
         return deferred { [weak self] () -> Observable<MoyaResponse> in